pyproject.toml
[tool.poetry]
name = "flask_template"
version = "0.1.0"
description = ""
authors = ["Gabriel Oger <gabriel.oger@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.3.3"
environs = "^9.5.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
factory-boy = "^3.3.0"
click = "^8.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
